+++ modules/client/src/main/java/cz/cvut/fit/baklaal1/client/handler/TeacherHandler.java
@@ -50,7 +50,7 @@ public class TeacherHandler extends PersonHandler<TeacherDTO, TeacherCreateDTO> 
 
         String username = args.getOptionValues(USERNAME).get(0);
         String name = args.getOptionValues(NAME).get(0);
-        Timestamp birthdate = args.getOptionValues(BIRTHDATE) != null ? new Timestamp(Integer.parseInt(args.getOptionValues(BIRTHDATE).get(0))) : null;
+        Timestamp birthdate = args.getOptionValues(BIRTHDATE) != null ? new Timestamp(Long.parseLong(args.getOptionValues(BIRTHDATE).get(0))) : null;
 
         String wageString = args.getOptionValues(WAGE) != null ? args.getOptionValues(WAGE).get(0) : "0";
         //TODO Number locales are hell
